requirements.txt
openai-whisper
moviepy
pydub
srt
